+++ lib/cfn.js
@@ -1,6 +1,8 @@
 const cf = require('@mapbox/cloudfriend');
 const path = require('path');
-const utils = require('./utils');
+const webhooks = require('./events/webhook');
+const snsEvent = require('./events/sns');
+const cloudWatchEvent = require('./events/cloudwatch');
 
 module.exports = buildFunctionTemplate; // returns full template
 module.exports.embed = buildFunction; // returns only lambda resources for template merge
@@ -184,7 +186,7 @@ function buildFunction(options) {
 
   // if no event source specified default to an SNS rule
   if (!options.eventSources || options.eventSources == {}) {
-    sns = buildSnsEvent(options);
+    sns = snsEvent.buildSnsEvent(options);
   }
 
   // only one of each event source type can be specified by a function
@@ -192,16 +194,16 @@ function buildFunction(options) {
     for (var event in options.eventSources) {
       switch(event) {
       case 'cloudwatchEvent':
-        cloudwatchEvent = buildCloudwatchEvent(options,'cloudwatchEvent');
+        cloudwatchEvent = cloudWatchEvent.buildCloudwatchEvent(options, 'cloudwatchEvent');
         break;
       case 'schedule':
-        schedule = buildCloudwatchEvent(options,'schedule');
+        schedule = cloudWatchEvent.buildCloudwatchEvent(options,'schedule');
         break;
       case 'sns':
-        sns = buildSnsEvent(options);
+        sns = snsEvent.buildSnsEvent(options);
         break;
       case 'webhook':
-        webhook = buildWebhookEvent(options);
+        webhook = webhooks.buildWebhookEvent(options);
         break;
       default:
         throw new Error('Unknown event source specified: ' + event);
@@ -384,355 +386,6 @@ function buildLambda(options) {
   }
 
   return fn;
-}
-
-module.exports.buildCloudwatchEvent = buildCloudwatchEvent;
-function buildCloudwatchEvent(options, functionType) {
-  if (!functionType) throw new Error('functionType property required for cloudwatch event');
-  if (!functionType.match(/cloudwatchEvent|schedule/)) throw new Error('unknown functionType property: ' + functionType);
-  if (functionType === 'cloudwatchEvent' && !options.eventSources.cloudwatchEvent.eventPattern) throw new Error('eventPattern required for cloudwatch event');
-  if (functionType === 'schedule' && !options.eventSources.schedule.expression) throw new Error('scheduled function expression cannot be undefined');
-
-  var eventName = options.name + utils.capitalizeFirst(functionType);
-  var event = {
-    Resources: {}
-  };
-
-  event.Resources[eventName + 'Permission'] = {
-    Type: 'AWS::Lambda::Permission',
-    Properties: {
-      FunctionName: cf.getAtt(options.name, 'Arn'),
-      Action: 'lambda:InvokeFunction',
-      Principal: 'events.amazonaws.com',
-      SourceArn: cf.getAtt(eventName + 'Rule', 'Arn')
-    }
-  };
-
-  event.Resources[eventName + 'Rule'] = {
-    Type: 'AWS::Events::Rule',
-    Properties: {
-      RoleArn: cf.getAtt('LambdaCfnRole', 'Arn'),
-      State: 'ENABLED',
-      Targets: [
-        {
-          Arn: cf.getAtt(options.name, 'Arn'),
-          Id: options.name
-        }
-      ]
-    }
-  };
-
-  if (functionType === 'cloudwatchEvent') {
-    event.Resources[eventName + 'Rule'].Properties.EventPattern = options.eventSources.cloudwatchEvent.eventPattern;
-  } else {
-    event.Resources[eventName + 'Rule'].Properties.ScheduleExpression = options.eventSources.schedule.expression;
-  }
-
-  return event;
-}
-
-module.exports.buildWebhookEvent = buildWebhookEvent;
-function buildWebhookEvent(options) {
-  var webhookName = options.name + 'Webhook';
-  var webhook = {
-    Resources: {},
-    Outputs: {}
-  };
-
-  if (!options.eventSources.webhook.method) {
-    throw new Error('Webhook function method not found');
-  }
-
-  if (/GET|HEAD|PUT|PATCH|OPTIONS|POST|DELETE/.test(options.eventSources.webhook.method.toUpperCase()) == false) {
-    throw new Error('Invalid client HTTP method specified: ' + options.eventSources.webhook.method);
-  }
-
-  webhook.Resources[webhookName + 'Resource'] = {
-    Type: 'AWS::ApiGateway::Resource',
-    Properties: {
-      ParentId: cf.getAtt(webhookName + 'ApiGateway', 'RootResourceId'),
-      RestApiId: cf.ref(webhookName + 'ApiGateway'),
-      PathPart: options.name.toLowerCase()
-    }
-  };
-
-  if (options.eventSources.webhook.method) {
-    if (!options.eventSources.webhook.methodResponses) {
-      options.eventSources.webhook.methodResponses = [
-        {
-          StatusCode: '200',
-          ResponseModels: {
-            'application/json': 'Empty'
-          }
-        },
-        {
-          StatusCode: '500',
-          ResponseModels: {
-            'application/json': 'Empty'
-          }
-        }
-      ];
-    } else if (!Array.isArray(options.eventSources.webhook.methodResponses)) {
-      throw new Error('Webhook method responses is not an array');
-    }
-    if (!options.eventSources.webhook.integration) {
-      options.eventSources.webhook.integration = {
-        Type: 'AWS',
-        IntegrationHttpMethod: 'POST',
-        Uri: cf.join(['arn:aws:apigateway:', cf.region, ':lambda:path/2015-03-31/functions/', cf.getAtt(options.name, 'Arn'), '/invocations']),
-        IntegrationResponses: [
-          {
-            StatusCode: '200'
-          },
-          {
-            StatusCode: '500',
-            SelectionPattern: '^(?i)(error|exception).*'
-          }
-        ]
-      };
-    } else {
-      // sets safe defaults for missing parameters if integration object is passed
-      options.eventSources.webhook.integration.Type = options.eventSources.webhook.integration.Type ? options.eventSources.webhook.integration.Type : 'AWS';
-      options.eventSources.webhook.integration.IntegrationHttpMethod = options.eventSources.webhook.integration.IntegrationHttpMethod ? options.eventSources.webhook.integration.IntegrationHttpMethod : 'POST';
-      options.eventSources.webhook.integration.Uri = options.eventSources.webhook.integration.Uri ? options.eventSources.webhook.integration.Uri : cf.join(['arn:aws:apigateway:', cf.region, ':lambda:path/2015-03-31/functions/', cf.getAtt(options.name, 'Arn'), '/invocations']);
-      options.eventSources.webhook.integration.IntegrationResponses = options.eventSources.webhook.integration.IntegrationResponses ? options.eventSources.webhook.integration.IntegrationResponses : [{ StatusCode: '200' }, { StatusCode: '500', SelectionPattern: '^(?i)(error|exception).*'}];
-    }
-
-    webhook.Resources[webhookName + 'Method'] = {
-      Type: 'AWS::ApiGateway::Method',
-      Properties: {
-        RestApiId: cf.ref(webhookName + 'ApiGateway'),
-        ResourceId: cf.ref(webhookName + 'Resource'),
-        AuthorizationType: 'None',
-        HttpMethod: options.eventSources.webhook.method.toUpperCase(),
-        MethodResponses: options.eventSources.webhook.methodResponses,
-        Integration: options.eventSources.webhook.integration
-      }
-    };
-  }
-
-  webhook.Resources[webhookName + 'Permission'] = {
-    Type: 'AWS::Lambda::Permission',
-    Properties: {
-      FunctionName: cf.getAtt(options.name, 'Arn'),
-      Action: 'lambda:InvokeFunction',
-      Principal: 'apigateway.amazonaws.com',
-      SourceArn: cf.join(['arn:aws:execute-api:', cf.region, ':', cf.accountId, ':', cf.ref(webhookName + 'ApiGateway'),'/*'])
-    }
-  };
-
-  webhook.Resources[webhookName + 'ApiGateway'] = {
-    Type: 'AWS::ApiGateway::RestApi',
-    Properties: {
-      Name: cf.stackName,
-      FailOnWarnings: 'true'
-    }
-  };
-
-  // randomizes deployment name so that code can be redeployed over an existing stage
-  var apiDeploymentRandom;
-  if (process.env.NODE_ENV == 'test') {
-    apiDeploymentRandom = 'ApiDeployment';
-  } else {
-    apiDeploymentRandom = 'ApiDeployment' + Math.random().toString(36).slice(2);
-  }
-
-  webhook.Resources[webhookName + 'ApiKey'] = {
-    Type: 'AWS::ApiGateway::ApiKey',
-    DependsOn: webhookName + apiDeploymentRandom,
-    Properties: {
-      Name: cf.stackName,
-      Enabled: 'true',
-      StageKeys: [
-        {
-          RestApiId: cf.ref(webhookName + 'ApiGateway'),
-          StageName: 'prod'
-        }
-      ]
-    }
-  };
-
-  webhook.Resources[webhookName + apiDeploymentRandom] = {
-    Type: 'AWS::ApiGateway::Deployment',
-    DependsOn: webhookName + 'Method',
-    Properties: {
-      RestApiId: cf.ref(webhookName + 'ApiGateway'),
-      StageName: 'prod'
-    }
-  };
-
-  webhook.Resources[webhookName + 'ApiLatencyAlarm'] = {
-    Type: 'AWS::CloudWatch::Alarm',
-    Properties: {
-      EvaluationPeriods: '5',
-      Statistic: 'Sum',
-      Threshold: '4',
-      AlarmDescription: 'https://github.com/mapbox/lambda-cfn/blob/master/alarms.md#ApiLatencyAlarm',
-      Period: '60',
-      AlarmActions: [
-        cf.ref('ServiceAlarmSNSTopic')
-      ],
-      Namespace: 'AWS/ApiGateway',
-      Dimensions: [
-        {
-          Name: 'APIName',
-          Value: cf.stackName
-        }
-      ],
-      ComparisonOperator: 'GreaterThanThreshold',
-      MetricName: 'Latency'
-    }
-  };
-
-  webhook.Resources[webhookName + 'Api4xxAlarm'] = {
-    Type: 'AWS::CloudWatch::Alarm',
-    Properties: {
-      EvaluationPeriods: '5',
-      Statistic: 'Sum',
-      Threshold: '100',
-      AlarmDescription: 'https://github.com/mapbox/lambda-cfn/blob/master/alarms.md#Api4xxAlarm',
-      Period: '60',
-      AlarmActions: [ cf.ref('ServiceAlarmSNSTopic') ],
-      Namespace: 'AWS/ApiGateway',
-      Dimensions: [
-        {
-          Name: 'APIName',
-          Value: cf.stackName
-        }
-      ],
-      ComparisonOperator: 'GreaterThanThreshold',
-      MetricName: '4xxError'
-    }
-  };
-
-  webhook.Resources[webhookName + 'ApiCountAlarm'] = {
-    Type: 'AWS::CloudWatch::Alarm',
-    Properties: {
-      EvaluationPeriods: '5',
-      Statistic: 'Sum',
-      Threshold: '10000',
-      AlarmDescription: 'https://github.com/mapbox/lambda-cfn/blob/master/alarms.md#ApiCountAlarm',
-      Period: '60',
-      AlarmActions: [
-        cf.ref('ServiceAlarmSNSTopic')
-      ],
-      Namespace: 'AWS/ApiGateway',
-      Dimensions: [
-        {
-          Name: 'APIName',
-          Value: cf.stackName
-        }
-      ],
-      ComparisonOperator: 'GreaterThanThreshold',
-      MetricName: 'Count'
-    }
-  };
-
-  webhook.Resources[webhookName + 'Permission'] = {
-    Type: 'AWS::Lambda::Permission',
-    Properties: {
-      FunctionName: cf.getAtt(options.name, 'Arn'),
-      Action: 'lambda:InvokeFunction',
-      Principal: 'apigateway.amazonaws.com',
-      SourceArn: cf.join(['arn:aws:execute-api:', cf.region, ':', cf.accountId, ':', cf.ref(webhookName + 'ApiGateway'), '/*'])
-    }
-  };
-
-  webhook.Outputs[webhookName + 'APIEndpoint'] = {
-    Value: cf.join(['https://', cf.ref(webhookName + 'ApiGateway'), '.execute-api.', cf.region, '.amazonaws.com/prod/', options.name.toLowerCase()])
-  };
-
-  if (options.eventSources.webhook.apiKey) {
-    webhook.Resources[webhookName + 'Method'].Properties.ApiKeyRequired = 'true';
-    webhook.Outputs[webhookName + 'ApiKey'] = {
-      Value: cf.ref(webhookName + 'ApiKey')
-    };
-  }
-
-  return webhook;
-}
-
-module.exports.buildSnsEvent = buildSnsEvent;
-function buildSnsEvent(options) {
-  var snsEvent = {
-    Resources: {},
-    Outputs: {}
-  };
-
-  snsEvent.Resources[options.name + 'SNSPermission'] = {
-    Type: 'AWS::Lambda::Permission',
-    Properties: {
-      FunctionName: cf.getAtt(options.name, 'Arn'),
-      Action: 'lambda:InvokeFunction',
-      Principal: 'sns.amazonaws.com',
-      SourceArn: cf.ref(options.name + 'SNSTopic')
-    }
-  };
-
-  snsEvent.Resources[options.name + 'SNSUser'] = {
-    Type: 'AWS::IAM::User',
-    Properties: {
-      Policies: [
-        {
-          PolicyName: options.name + 'SNSTopicPolicy',
-          PolicyDocument: {
-            Version: '2012-10-17',
-            Statement: [
-              {
-                Resource: cf.ref(options.name + 'SNSTopic'),
-                Action: [
-                  'sns:ListTopics',
-                  'sns:Publish',
-                ],
-                Effect: 'Allow'
-              },
-              // required permissions for Zapier SNS integrations
-              {
-                Resource: cf.join(['arn:aws:sns:', cf.region, ':', cf.accountId, ':*']),
-                Action: [
-                  'sns:ListTopics',
-                ],
-                Effect: 'Allow'
-              }
-            ]
-          }
-        }
-      ]
-    }
-  };
-
-  snsEvent.Resources[options.name + 'SNSTopic'] = {
-    Type: 'AWS::SNS::Topic',
-    Properties: {
-      DisplayName: cf.join('-', [cf.stackName, options.name]),
-      TopicName: cf.join('-', [cf.stackName, options.name]),
-      Subscription: [
-        {
-          Endpoint: cf.getAtt(options.name, 'Arn'),
-          Protocol: 'lambda'
-        }
-      ]
-    }
-  };
-
-  snsEvent.Resources[options.name + 'SNSUserAccessKey'] = {
-    Type: 'AWS::IAM::AccessKey',
-    Properties: {
-      UserName: cf.ref(options.name + 'SNSUser')
-    }
-  };
-
-  snsEvent.Outputs[options.name + 'SNSTopic'] = {
-    Value: cf.ref(options.name + 'SNSTopic')
-  };
-  snsEvent.Outputs[options.name + 'SNSUserAccessKey'] = {
-    Value: cf.ref(options.name + 'SNSUserAccessKey')
-  };
-  snsEvent.Outputs[options.name + 'SNSUserSecretAccessKey'] = {
-    Value: cf.getAtt(options.name + 'SNSUserAccessKey', 'SecretAccessKey')
-  };
-
-  return snsEvent;
 }
 
 module.exports.buildRole = buildRole;
@@ -999,3 +652,7 @@ function buildSnsDestination(options) {
     return sns;
   }
 }
+
+module.exports.buildWebhookEvent = webhooks.buildWebhookEvent;
+module.exports.buildSnsEvent = webhooks.buildSnsEvent;
+module.exports.buildCloudwatchEvent = cloudWatchEvent.buildCloudwatchEvent;
